requirements.txt
openai-whisper
streamlit
librosa
pytest
pytube